mononoke/filestore: introduce peek()
Summary:
This adds a new `peek(.., N, ..)` call in the Filestore that allows reading at least N bytes from a file in the Filestore. This is helpful for generating Mercurial metadata blobs.

This is implemented using the same ChunkStream we use to write to the Filestore (but that needed a little fixing to support streams of empty bytes as well).

Reviewed By: aslpavel

Differential Revision: D16440672

fbshipit-source-id: a584099f87ab34e2151b9c3f5c9f1289575f024b