mononoke/filestore: make reads faster
Summary: As noted earlier in this stack, the Filestore read implementation was very inefficient, since it required reading a Chunk before moving on to the next one. Since our Chunked files will typically have just one level of chunking, this is very inefficient (we could be fetching additional chunks ahead of time). This new implementation lets us take advantage of buffering, so we can load arbitrarily as many chunks in parallel as we'd like.

Reviewed By: aslpavel

Differential Revision: D16560767

fbshipit-source-id: c02c10c5de0fc5fdc3ee3897ae855b316ea34605