mononoke/filestore: switch to functions (config is only needed in writes)
Summary: This updates the Filestore API to be a set of functions, instead of a Struct. The rationale here is that there is only one Filestore call that needs anything on top of a blobstore (the write call), so making those functions makes it much easier to incorporate Filestore in various places that need it without having to thread a Filestore all the way through.

Reviewed By: aslpavel

Differential Revision: D16440668

fbshipit-source-id: 4b4bc8872e205a66a12ec96a478f0f1811f2e6b1