mononoke: move typed fetch to Blobstore Trait
Summary:
This moves typed fetches from the blobrepo to the blobstore. The upshot is that this allows consumers of a blobstore to do typed fetches, instead of forcing them to get bytes then cast bytes to a blob, then cast the blob to the thing they want.

This required refactoring our crate hierarchy a little bit by moving BlobstoreBytes into Mononoke Types, since we now need the Blobstore crate to depend on Mononoke types, whereas it was the other way around before (since BlobstoreValue was already there, that seems reasonable)

Reviewed By: StanislavGlebik

Differential Revision: D16486774

fbshipit-source-id: 05751986ce3cb7273d68a8b4ebe9957bb892bcd6