mononoke/filestore: dont boxify
Summary: This updates the Filestore to avoid boxifying in its chunking code. The upshot is that this gets us to a place where passing a Send Stream into the Filestore gives you a Send Future back, and passing in a non-Send Stream in the Filestore gives you a non-Send future back (I hinted at this earlier in the diff that introduced faster writes).

Reviewed By: aslpavel

Differential Revision: D16560768

fbshipit-source-id: b77766380f2eaed5919f78cef6fbc02afeead0b9