mononoke/filestore: Make ContentMetadata contents mandatory
Summary:
NOTE: This makes Thrift changes, but we never stored anything in this format.

This updates `ContentAliasBackmapping` to make all fields we are currently computing mandatory. We never stored anything using the old format, so this is safe to change for now.

In general, this makes code a little simpler, since we can rely on those fields to exist. Later on, if we add new backmappings, then we'll have to handle cases where they might not exist, but there's no reason to force this work upon ourselves for now if we don't need to.

Reviewed By: aslpavel

Differential Revision: D16440673

fbshipit-source-id: 6f5d0e4a687a2641a9b5e8e518859b796997e22c