mononoke/filestore: use a strongly typed ExpectedSize to avoid confusing hints and real sizes
Summary: This updates the Filestore implementation to use a ExpectedSize type to differentiate sizes we know to be correct from sizes that were given to us by writers. This ensures we can't accidentally use the ExpectedSize we received from the writer when we should be using the effective observed size.

Reviewed By: aslpavel

Differential Revision: D16440674

fbshipit-source-id: 8bf03b9a962339ea2896f2f60d4a52417ca0327e