lfs/http keepalive: fix broken uploads on retry
Summary:
The KeepAlive HTTP implementation is bugged in it’s retry logic, it supports reading from a file pointer, but doesn’t support rewinding of the seek cursor when it performa a retry. So it can happen that an upload fails for whatever reason and will then ‘hang’ on the retry event.

The sequence of events that get triggered are:
 - Upload file A, goes OK. Keep-Alive caches connection.
 - Upload file B, fails due to (for example) failing Keep-Alive, but LFS file pointer has
   been consumed for the upload and fd has been closed.
 - Retry for file B starts, sets the Content-Length properly to the expected file size, but since file pointer has been consumed no data will be uploaded, causing the server to wait for the uploaded data until either client or server reaches a timeout, making it seem as our mercurial process hangs.

This is just a stop-gap measure to prevent this behavior from blocking Mercurial (LFS has retry logic). A proper solutions need to be build on top of this stop-gap measure: for upload from file pointers, we should support fseek() on the interface,
Since we expect to consume the whole file always anyways, this should bafe. This way we can seek back to the beginning on a retry.

Reviewed By: quark-zju

Differential Revision: D16604844

fbshipit-source-id: 9a0d418f00520bda499272b864ff24c99e888aa2