lfs: rewind fp on retries
Summary:
To avoid the scenario where we are incorrectly uploading the data that we're
supposed to upload, we need to reset the file pointer to read from the
beginning so that for each retry of uploading we get the correct data.

Previously before this patch, Mercurial would have nothing to upload on a retry
and the connection would 'hang' waiting for the data to be send to the server.

Reviewed By: krallin

Differential Revision: D16620072

fbshipit-source-id: 2238b99ff612e06dc0640717bc42fe345c35037d