remotefilelog: remove old files in the hgcache that aren't packfiles
Summary:
The hgcache packs directory is being written to in several places, most of the
files being created there are temporary ones and renamed once ready to be
visible. In some cases, when hg crashes midway, or when the packfile is mmap'ed
on windows, temporary files are left around.

While we were trying to remove these files before, the solution taken involved
removing files we knew were safe to be removed. Unfortunately, that meant we
missed some of the temporary files. Let's instead switch to removing everything
but packfiles and the repacklock.

Reviewed By: kulshrax

Differential Revision: D16657105

fbshipit-source-id: f4dd45aa71c02acba52a745b9bec3019d20e03c4