manifest: update manifest::get to return optional enum
Summary:
This allows us to say that the queried path was a directory and simplifies
the API a little bit.

The Python/C++ Manifest API provide inspection of directories through several
functions. We want the Rust manifest to provide the same functionality.
This approach looks to have more consistency as the API evolves.

`FsNode` is a structure that is forward looking. We will want to add an API
that lists a directory and an interator of FsNode fits that well.
I also felt the need for it over the development of the manifest code.

Reviewed By: quark-zju

Differential Revision: D16609253

fbshipit-source-id: f826d7b21e3001f4bef43a35b9d1a9bc5a59eda9