remotefilelog: start HTTP file fetching progress bars at 0
Summary:
Previously, we would use `None` as the start value for HTTP data fetching progress bars so that the bar would be rendered as a spinner until we received a response (which would contain a content size header, allowing us to set the progress bar total).

This would result in incorrect behavior if the server took a long time to return a response, because by default if no start value is provided the progress bar framework will display the number of seconds that have elapsed. Given that we've specified a byte count format function for the progress value, the progress bar would end up displaying the number of seconds elapsed as a byte count, which was very confusing.

Reviewed By: quark-zju

Differential Revision: D16663124

fbshipit-source-id: 0e224b3b670abc803356214cd12205a6f4259cd9