revisionstore: add a basic packstore in Rust
Summary:
The PackStore is one of the last missing piece to be able to have all the code
that deals with data on disk in Rust. There are a couple of features that the
Python code has that isn't present here:
 1) LRU of packfiles
 2) Detection of corrupted packfiles
 3) Automatic repacking

While 1) and 2) will be added later, I'm not planning on working on 3). The
only reason we still have 3) is due to hg_memcache_client potentially writing
histpack one by one. I'm planning on adding a very basic IndexedLog based
history store and have hg_memcache_client write to it. This will greatly reduce
the number of histpack created and thus remove the need for the repacks.

Reviewed By: sfilipco

Differential Revision: D16606429

fbshipit-source-id: 06bcdad6b79de4cd6284c685cccf6fdf7ab3f8a7