fix hg svn info
Summary:
The logic is currently a bit broken in a weird combination:

The current configurations are
- usefirstpublicancestorinsvninfo=False
- useglobalrevinsvninfo=True

This means hg will have to traverse back to the last hg commit with svnrev. It will get
[slower and slower](https://fburl.com/scuba/skfj8p7u) to find those commits as we are having more commits without snvrev.

Ideally we can turn `usefirstpublicancestorinsvninfo` to True and that will then only traverse back to the last public ancestor. However it will trigger another bug because this public ancestor won't be in the old svn revmap and therefore hg will just bail with "Not a child of an svn revision".

Therefore the fix is to fix that bug by moving the `if pn not in hashes:` into the else part of the config and also get the subdir value from config rather than from the `extras['convert_info']`.

The next step is to set
- usefirstpublicancestorinsvninfo=True
- useglobalrevinsvninfo=True
- usereposubdirinsvninfo=True
once this diff is deployed to stable.

Reviewed By: DurhamG

Differential Revision: D16717554

fbshipit-source-id: c05b9882aa9ddeeeefd9315b478ba085fbc9bae7