Add method to find all path component manifest entries
Summary: Building on the previous diff, expose the newly added ability to fetch the manifest entries of all path components of the given set of paths as a new method on BlobRepo. This method will be called from the API server as a way to perform selective prefetching of manifest nodes for client commands that operate on only a subset of paths for a given commit.

Reviewed By: krallin

Differential Revision: D16704978

fbshipit-source-id: 6ab68ff7c0d0b76d3317d8e2bd6d9085f52f3477