Modify QueryTree to allow selecting all path components
Summary:
The `find_entries_in_manifest` method uses a data structure called
`QueryTree` to guide manifest traversal. Previously, this traversal would
only return the manifest entries corresponding to the final component of each
path traversed. This diff adds the optional ability to return the entry for
every path component traversed, regardless of whether that entry was explicitly
requested. This is useful for prefetching in Mercurial, where we'd like to
cache these entries locally on the client to avoid unnecessary server
roundtrips when accessing trees with common path prefixes.

Reviewed By: krallin

Differential Revision: D16704979

fbshipit-source-id: 6436c1f8772cc4a8db74994435be8ba87352ff44