mononoke: refactor blacklist add admin command to extract boilerplate
Summary:
This is done to make `remove` and `list` be implementable in terms of the same
boilerplate with some different closures.

NB: I don't particularly like how this refactoring looks, but I do think it
looks better than copy/pasting the code and I think it is of no major
importance, given that this is an admin command.

Reviewed By: krallin

Differential Revision: D16643437

fbshipit-source-id: 7f64b112eda1ce963dffed897bf2c9fa42234d79