redaction: handle redacted content coming from the EdenAPI
Summary:
In some cases, the file we are trying to fetch from the EdenAPI can be
redacted. We need to handle those cases. The good way to do so would be to
support this on a protocol level, but that is a *lot* of work. For now,
we are using the "tombstone approach", in which the ApiServer reacts to
redaction errors by injecting a special "tomstone" content, which the client
needs to recognize as not an error, but as a special case. In particular,
it needs to not fail the validation.

Reviewed By: markbt

Differential Revision: D16752747

fbshipit-source-id: 1bf126d50ecd6b862954284461b7db8f51e7e891