rotatelog: make indexes of immutable Logs up-to-date
Summary:
The indexes are lagging by design to reduce space overhead. For immutable Logs
(esp. used by RotateLog), the indexes no longer need to be lagging.

Make it so to reduce overhead opening RotateLog.

As we're here, also fix an issue where `rotate_assume_locked` was not really
protected by a lock.

Reviewed By: sfilipco

Differential Revision: D16587181

fbshipit-source-id: 3cf81864e90c875ee661dbd994bcd3ebc4b55322