eden debug file_stat - add --sizes arg to tool
Summary:
eden debug file_stat - add --sizes arg to tool
Currently using naive shell call and temprary file to store all of the paths to
get the sizes. This is significantly slower than running without sizes and will
be replaced with proper sizes provided by eden thrif call in the future.

This is sufficient to establish the structure of the output of this command and
start integrating it with sandcastle

Reviewed By: strager

Differential Revision: D16836053

fbshipit-source-id: ad47416a0036d1046768fa9b9b6c9dfefbb8edd4