mononoke/blobstore: remove typed fetch, and use Loadable instead
Summary:
This updates an earlier diff I had that introduced the typed fetch method on the Blobstore trait. Since we now have Storable and Loadable traits, it makes a little more sense to implement those for MononokeId and use that instead.

I was hoping this would mean we could move the implementation of those traits into the mononoke_types crate, but unfortunately it looks like the `where T: MononokeId` implementation for `Loadable` can't live there (Rust wants it next to the trait definition).

That said, this change does mean we can get rid of some manual BlobstoreBytes creation here and there, so that's still an upside.

In doing so, I updated the Loadable and Storable traits to require only a Blobstore reference, and to take ownership of what they're storing. This saves some clones, especially in the latter.

Reviewed By: aslpavel

Differential Revision: D16733733

fbshipit-source-id: 3ec3a97cef90d6850f2939ba3f5ec6e6e5b459fb