mononoke: Add rechunking command to filestore
Summary:
Once enabled, the filestore will chunk up new files before storing them
in the blobstore. However, there are quite a few large files in the blobstore that should
be rechunked.

Add support to the filestore for rechunking files by downloading them from the
blobstore and reuploading the, but chunked.

Reviewed By: krallin

Differential Revision: D16802702

fbshipit-source-id: 2e617ec2cd215204596769c72431fd921ff33dd8